LU-4856 misc: 64 bit counters for objects tied to memory size
When the number of an object in use or circulation is tied to memory
size of the system, very large memory systems can overflow 32 bit
counters.  This patch converts several of the most obviously exposed
counters - those initialized on the basis of memory size.
This incudes the debug log size, client osc pool/lru, and ldlm pool.
Several additional checks are made to detect overflow on variables
which have not been converted.

+++ libcfs/libcfs/tracefile.c
@@ -1112,7 +1112,7 @@ void cfs_trace_stop_thread(void)
 	mutex_unlock(&cfs_trace_thread_mutex);
 }
 
-int cfs_tracefile_init(int max_pages)
+int cfs_tracefile_init(long max_pages)
 {
 	struct cfs_trace_cpu_data *tcd;
 	int	i;
